split methods into two classes (0509) #30
